build: add static-analysis + UBSan tooling and fix surfaced findings
Wire up three runtime/static checkers and clean the tree for each.

Tooling:
- clang-tidy: .clang-tidy (analyzer + bugprone, tuned for Asterisk C)
  plus `make tidy` / `make compile-commands` (bear-recorded DB so the
  pjproject config_site.h overlay is in scope) and a non-blocking CI
  step on the 22 cell.
- cppcheck: `make cppcheck` (second-opinion engine; third-party headers
  fed as -I so macros resolve, their findings suppressed) + CI step.
- sanitize-ub: UBSan build variant into OBJ_DIR/sanitize-ub (isolated
  so it can't be installed over production), documented bench recipe in
  tests/README.md.

Build hardening:
- enable -Wconversion / -Wsign-conversion under -Werror; move asterisk
  and pjproject includes to -isystem so their unfixable conversion
  noise (~440 warnings) is suppressed while ours still surface.

Findings fixed:
- header guards: drop leading underscore (_RES_PJSIP_CISCO_*_H reserved).
- atoi -> ast_str_to_int in conference applicationid / ConfList parsing.
- make implicit string compares explicit (strcmp(...) != 0).
- explicit casts for the size_t/int/char/pj_ssize_t conversions and the
  ast_presence_state enum<->int signedness clang flags.

Build clean under -Werror; `make tidy` and `make cppcheck` report zero
findings against asterisk 22.9.0.

+++ res/cisco_endpoint/endpoint.c
@@ -105,7 +105,7 @@ const char *cisco_cfwd_get(const char *endpoint_id, char *buf, size_t buflen)
 	if (ast_strlen_zero(endpoint_id)) {
 		return buf;
 	}
-	if (ast_db_get("CF", endpoint_id, buf, buflen)) {
+	if (ast_db_get("CF", endpoint_id, buf, (int) buflen)) {
 		buf[0] = '\0';
 	}
 	return buf;
